Refactor test a bit.
